requirements.txt
python-dotenv>=1.0.0
click>=8.1.7
loguru>=0.7.2
zipfile38>=0.0.3
requests>=2.31.0